Corrige error

+++ src/models/MongoModel.ts
@@ -1,5 +1,5 @@
-import { Model, /* isValidObjectId */ UpdateQuery } from 'mongoose';
-/* import { ErrorTypes } from '../errors/catalog'; */
+import { Model, isValidObjectId, UpdateQuery } from 'mongoose';
+import { ErrorTypes } from '../errors/catalog';
 import { IModel } from '../interfaces/IModel';
 
 abstract class MongoModel<T> implements IModel<T> {
@@ -18,12 +18,15 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async readOne(_id: string): Promise<T | null> {
-    /*  if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidMongoId); */
+    if (_id.length < 24) throw new Error(ErrorTypes.FewCharacters);
+
+    if (Number(_id) && _id.length === 24) throw Error(ErrorTypes.InvalidId);
+
     return this._model.findOne({ _id });
   }
 
   public async update(_id: string, obj: T): Promise<T | null> {
-    /*  if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidMongoId); */
+    if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidId);
     return this._model.findByIdAndUpdate(
       { _id },
       { ...obj } as UpdateQuery<T>,
@@ -32,7 +35,7 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async delete(_id: string): Promise<T | null> {
-    /*  if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidMongoId); */
+    if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidId);
     return this._model.findByIdAndDelete({ _id });
   }
 }
